Feat: 이미지 미리보기, 예외처리, 사이드바 구현

+++ src/pages/GalleryDetail.tsx
@@ -1,8 +1,9 @@
-import { useState, useEffect } from "react"
+import { useState, useEffect, useContext } from "react"
 import styled from "styled-components"
 import { db } from "../firebase"
 import { doc, collection, getDocs, deleteDoc } from  "firebase/firestore"
 import { Link, useParams, useNavigate } from "react-router-dom";
+import { AuthContext } from "authentication/authContext";
 
   // 함수 인자 타입 선언
   interface GalleryDetailProps {
@@ -25,6 +26,7 @@ const GalleryDetail: React.FC<GalleryDetailProps> = ({setOnEdit}) => {
   // 현재 url의 id값 구하기
   const { id } = useParams<string>();
   const navigate = useNavigate();
+  const user = useContext(AuthContext);
 
   const [users, setUsers] = useState<any[]>([]);
   const usersCollectionRef = collection(db, "gallery");
@@ -43,17 +45,53 @@ const GalleryDetail: React.FC<GalleryDetailProps> = ({setOnEdit}) => {
   
   // 데이터 삭제하기
   const deleteGallery = async (id: string) => {
+    if(user){
+    const userToDelete = users.find(user => user.id === id);
+    if(userToDelete && user.uid === userToDelete.uid){
     // 삭제 여부 확인
-    const confirmDelete = window.confirm("정말로 삭제하시겠습니까?")
-    if(confirmDelete){
-      // 삭제할 id의 데이터 지우기
-      const userDoc = doc(db, "gallery", id);
-      await deleteDoc(userDoc)
-      alert('삭제 완료했습니다')
-      // 삭제 후 리스트페이지로 이동
-      navigate('/Gallery')
-    } else{
+    const confirmDelete = window.confirm("정말 삭제하시겠습니까?")
+      if(confirmDelete){
+        // 삭제할 id의 데이터 지우기
+        const userDoc = doc(db, "gallery", id);
+        await deleteDoc(userDoc)
+        alert('삭제 완료했습니다')
+        // 삭제 후 리스트페이지로 이동
+        navigate('/Gallery')
+      } else{
+        return;
+      }
+    } else {
+      alert('해당 작성자만 삭제 가능합니다')
       return;
+    }
+    } else {
+      const confirmed = window.confirm("로그인해야 이용할 수 있습니다. 로그인 하시겠습니까?")
+      if (confirmed) {
+        navigate('/login')
+      } else {
+        return;
+      }
+    }
+  }
+
+  // 데이터 수정하기
+  const editGallery = async (id: string) => {
+    if(user){
+    const userToEdit = users.find(user => user.id === id);
+    if(userToEdit && user.uid === userToEdit.uid){
+      setOnEdit(true)
+      navigate(`/Gallery/edit/${id}`)
+    } else {
+      alert('해당 작성자만 수정 가능합니다')
+      return;
+    }
+    } else {
+      const confirmed = window.confirm("로그인해야 이용할 수 있습니다. 로그인 하시겠습니까?")
+      if (confirmed) {
+        navigate('/login')
+      } else {
+        return;
+      }
     }
   }
 
@@ -68,9 +106,7 @@ const GalleryDetail: React.FC<GalleryDetailProps> = ({setOnEdit}) => {
               </div>
               <div className="Gallery__btn-wrap">
               <button onClick={() => {deleteGallery(user.id)}} className="Gallery__btn delete">삭제</button>
-              <Link to={`/Gallery/edit/${user.id}`}>
-                <button onClick={() => setOnEdit(true)} className="Gallery__btn">수정</button>
-              </Link>
+              <button onClick={() => {editGallery(user.id)}} className="Gallery__btn">수정</button>
               </div>
             </GalleryHeader>
             <GalleryDesc>
@@ -84,6 +120,11 @@ const GalleryDetail: React.FC<GalleryDetailProps> = ({setOnEdit}) => {
         )
       })
     }
+        <GalleryBtn>
+        <Link to="/Gallery">
+          <button>목록으로</button>
+        </Link>
+        </GalleryBtn>
     </>
   )
 }
@@ -146,6 +187,21 @@ const GalleryEditor = styled.div`
   border-top: 1px solid #ddd;
   margin-top:20px;
   padding-top:10px;
+`
+
+const GalleryBtn = styled.div`
+margin-top: 20px;
+text-align:center;
+  button{
+    background-color: var(--main-color);
+    width: 95px;
+    height: 35px;
+    color: white;
+    border-radius: 4px;
+    border: none;
+    outline: none;
+    cursor: pointer;
+  }
 `
 
 export default GalleryDetail
